+++ apps/Rhino/COMPAS/dev/COMPAS_test_mesh_cmd.py
@@ -1,0 +1,26 @@
+from __future__ import print_function
+from __future__ import absolute_import
+from __future__ import division
+
+import compas
+from compas.datastructures import Mesh
+from compas_ui.app import App
+
+
+__commandname__ = 'COMPAS_test_mesh'
+
+
+def RunCommand(is_interactive):
+
+    app = App(name='COMPAS')
+
+    mesh = Mesh.from_obj(compas.get('tubemesh.obj'))
+    mesh.name = 'TubeMesh'
+
+    app.scene.add(mesh)
+    app.scene.update()
+    app.record()
+
+
+if __name__ == '__main__':
+    RunCommand(True)