+++ src/compas_ui/objects/__init__.py
@@ -1,18 +1,50 @@
+"""
+********************************************************************************
+objects
+********************************************************************************
+
+.. currentmodule:: compas_ui.objects
+
+
+Classes
+=======
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    Object
+    MeshObject
+    NetworkObject
+    VolMeshObject
+
+
+Exceptions
+==========
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    ObjectNotRegistered
+
+"""
+
 from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
 
-from compas.datastructures import Mesh
-
-from .exceptions import DataObjectNotRegistered
+from .exceptions import ObjectNotRegistered
 from .object import Object
 from .meshobject import MeshObject
-
-Object.register(Mesh, MeshObject, context='Rhino')
+from .networkobject import NetworkObject
+from .volmeshobject import VolMeshObject
 
 
 __all__ = [
-    'DataObjectNotRegistered',
+    'ObjectNotRegistered',
     'Object',
-    'MeshObject'
+    'MeshObject',
+    'NetworkObject',
+    'VolMeshObject'
 ]